requirements.txt
pyrogram
tgcrypto
tensorflow-cpu
tensorflow-hub
tf-keras
pillow
numpy
